+++ tests/unit_stream_mixin.cu
@@ -47,18 +47,18 @@ TEST(smallgwn_unit_stream_mixin, swap_exchanges_streams) {
 // ---------------------------------------------------------------------------
 
 TEST_F(CudaStreamFixture, geometry_object_default_stream) {
-    gwn::gwn_geometry_object<float, std::int64_t> geometry;
+    gwn::gwn_geometry_object<float, std::uint32_t> geometry;
     EXPECT_EQ(geometry.stream(), gwn::gwn_default_stream());
 }
 
 TEST_F(CudaStreamFixture, geometry_object_set_stream) {
-    gwn::gwn_geometry_object<float, std::int64_t> geometry;
+    gwn::gwn_geometry_object<float, std::uint32_t> geometry;
     geometry.set_stream(stream_a_);
     EXPECT_EQ(geometry.stream(), stream_a_);
 }
 
 TEST_F(CudaStreamFixture, geometry_object_clear_rebinds_stream) {
-    gwn::gwn_geometry_object<float, std::int64_t> geometry;
+    gwn::gwn_geometry_object<float, std::uint32_t> geometry;
     geometry.set_stream(stream_a_);
     EXPECT_EQ(geometry.stream(), stream_a_);
 
@@ -71,18 +71,18 @@ TEST_F(CudaStreamFixture, geometry_object_clear_rebinds_stream) {
 // ---------------------------------------------------------------------------
 
 TEST_F(CudaStreamFixture, bvh_object_default_stream) {
-    gwn::gwn_bvh_object<float, std::int64_t> bvh;
+    gwn::gwn_bvh_object<float, std::uint32_t> bvh;
     EXPECT_EQ(bvh.stream(), gwn::gwn_default_stream());
 }
 
 TEST_F(CudaStreamFixture, bvh_object_set_stream) {
-    gwn::gwn_bvh_object<float, std::int64_t> bvh;
+    gwn::gwn_bvh_object<float, std::uint32_t> bvh;
     bvh.set_stream(stream_a_);
     EXPECT_EQ(bvh.stream(), stream_a_);
 }
 
 TEST_F(CudaStreamFixture, bvh_object_clear_rebinds_stream) {
-    gwn::gwn_bvh_object<float, std::int64_t> bvh;
+    gwn::gwn_bvh_object<float, std::uint32_t> bvh;
     bvh.set_stream(stream_a_);
     bvh.clear(stream_b_);
     EXPECT_EQ(bvh.stream(), stream_b_);
@@ -93,18 +93,18 @@ TEST_F(CudaStreamFixture, bvh_object_clear_rebinds_stream) {
 // ---------------------------------------------------------------------------
 
 TEST_F(CudaStreamFixture, data_object_default_stream) {
-    gwn::gwn_bvh_moment_object<float, std::int64_t> data;
+    gwn::gwn_bvh_moment_object<float, std::uint32_t> data;
     EXPECT_EQ(data.stream(), gwn::gwn_default_stream());
 }
 
 TEST_F(CudaStreamFixture, data_object_set_stream) {
-    gwn::gwn_bvh_moment_object<float, std::int64_t> data;
+    gwn::gwn_bvh_moment_object<float, std::uint32_t> data;
     data.set_stream(stream_a_);
     EXPECT_EQ(data.stream(), stream_a_);
 }
 
 TEST_F(CudaStreamFixture, data_object_clear_rebinds_stream) {
-    gwn::gwn_bvh_moment_object<float, std::int64_t> data;
+    gwn::gwn_bvh_moment_object<float, std::uint32_t> data;
     data.set_stream(stream_a_);
     data.clear(stream_b_);
     EXPECT_EQ(data.stream(), stream_b_);
@@ -115,22 +115,22 @@ TEST_F(CudaStreamFixture, data_object_clear_rebinds_stream) {
 // ---------------------------------------------------------------------------
 
 TEST_F(CudaStreamFixture, geometry_move_construct) {
-    gwn::gwn_geometry_object<float, std::int64_t> src;
+    gwn::gwn_geometry_object<float, std::uint32_t> src;
     src.set_stream(stream_a_);
-    gwn::gwn_geometry_object<float, std::int64_t> dst(std::move(src));
+    gwn::gwn_geometry_object<float, std::uint32_t> dst(std::move(src));
     EXPECT_EQ(dst.stream(), stream_a_);
 }
 
 TEST_F(CudaStreamFixture, bvh_move_construct) {
-    gwn::gwn_bvh_object<float, std::int64_t> src;
+    gwn::gwn_bvh_object<float, std::uint32_t> src;
     src.set_stream(stream_a_);
-    gwn::gwn_bvh_object<float, std::int64_t> dst(std::move(src));
+    gwn::gwn_bvh_object<float, std::uint32_t> dst(std::move(src));
     EXPECT_EQ(dst.stream(), stream_a_);
 }
 
 TEST_F(CudaStreamFixture, data_move_construct) {
-    gwn::gwn_bvh_moment_object<float, std::int64_t> src;
+    gwn::gwn_bvh_moment_object<float, std::uint32_t> src;
     src.set_stream(stream_a_);
-    gwn::gwn_bvh_moment_object<float, std::int64_t> dst(std::move(src));
+    gwn::gwn_bvh_moment_object<float, std::uint32_t> dst(std::move(src));
     EXPECT_EQ(dst.stream(), stream_a_);
 }
